pyproject.toml
[tool.poetry]
name = "kuma"
version = "0.0.0"
license = "MPL-2.0"
description = "The MDN Web Docs site"
authors = ["MDN Devs <mdn-dev@mozilla.com>"]
homepage = "https://developer.mozilla.org"
repository = "https://github.com/mdn/kuma"
documentation = "https://kuma.readthedocs.io"

[tool.poetry.dependencies]
python = "^3.8"

babel = "^2.8.0"
boto3 = "^1.10.37"
botocore = "^1.14.17"
celery = "^4.3.0"
dj-database-url = "^0.5.0"
dj-email-url = "^1"
django = "^2"
django-allauth = "^0.41.0"
django-decorator-include = "^2.1"
django-extensions = "^2.2.6"
django-jinja = "^2.7.0"
django-mysql = "^3.3.0"
django-pipeline = "^1.7.0"
django-ratelimit = "2.0.0"
django-redis = "^4.12.1" # (Django cache backend)
django-statici18n = "^1.9.0"
django-storages = "^1.9.1"
django-taggit = "^0.24.0"
django-waffle = "^0.20.0"
djangorestframework = "^3.11.0"
elasticsearch = "^7.10.0"
elasticsearch-dsl = "^7.3.0"
enmerkar = "^0.7.1" # Maintained fork of django-babel
google-api-python-client = "^1.7.11"
gunicorn = "^20.1.0"
html5lib = "1.0.1"
httplib2 = "^0.17.0"
jinja2 = "^2.11.3"
lxml = "^4.6.3"
meinheld = "1.0.1"
mysqlclient = "^1.4.6" # (Django database driver)
newrelic = "^5.14.0.142"
oauth2client = "^4.1.3"
polib = "1.1.0"
puente = "0.5.0"
pyquery = "1.4.1"
python-dateutil = "^2.8.0"
python-decouple = "^3.3"
python-magic = "0.4.15"
pytz = "^2019.3"
raven = "6.10.0"
redo = "^2.0.3"
requests = "^2.25.1"
stripe = "^2.42.0"
urllib3 = "^1.25.11"
urlobject = "^2.4.3"
urlwait = "^1.0"
gevent = {extras = ["gevent"], version = "^20.5.2"}
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]
# Development Tools
django-debug-toolbar = "^2.2"
django-querycount = "^0.7.0"
werkzeug = "^1.0" # Enables runserver_plus from django-extensions

# Testing
braceexpand = "^0.1.6"
pytest = "~5.3"
pytest-base-url = "^1.4.1"
pytest-cov = "~2.8.1"
pytest-django = "~3.8.0"
pytest-metadata = "^1.11.0"
pytest-rerunfailures = "^8.0"
requests-mock = "^1.9.2"

ElasticMock = "^1.8.0"

# Linting
black = "20.8b1"
flake8 = "^3.8.1"
flake8-import-order = "^0.18.1"
dennis = "^0.9" # Used by `make localetest` to lint po files

# Pinned Dependencies
coverage = {extras = ["toml"], version = "^5"} # Use optional toml support

[tool.black]
target-version = ["py38"]

[tool.coverage.run]
source = ["kuma"]
branch = true
dynamic_context = "test_function"

[tool.coverage.report]
omit = ["*migrations*", "*/management/commands/*"]

[tool.coverage.html]
show_contexts = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
